Abstract ready logic for a particular element into a utility. Use that in all lifecycle handlers.
