+++ src/api/detached.js
@@ -1,9 +1,10 @@
 import apiWatch from './watch';
+import utilElementReady from '../util/element-ready';
 
 export default function (elem, name, func) {
   apiWatch(elem, function (added, removed) {
     for (let a = 0; a < removed.length; a++) {
-      func(removed[a]);
+      utilElementReady(removed[a], name, func);
     }
   });
 }
